% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animal_function.r
\name{animal_function}
\alias{animal_function}
\title{Animal Function}
\usage{
animal_function(love = TRUE)
}
\arguments{
\item{love}{Do you love animals? Defaults to TRUE.}
}
\description{
This function allows you to express your love of animals.
}
\examples{
animal_function()
}
\keyword{animal}
