% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_chr1.R
\docType{data}
\name{genome_chr1}
\alias{genome_chr1}
\title{First chromosome data from the 1000 Genomes Project}
\format{
\code{genome_chr1} is a 1092 x 50000 matrix.
}
\source{
\url{https://www.internationalgenome.org/}
}
\description{
The 1000 Genomes Project created a catalogue of common human
genetic variation, using openly consented samples from people who
declared themselves to be healthy. The reference data resources
generated by the project remain heavily used by the biomedical
science community. The International Genome Sample Resource (IGSR)
maintains and shares the human genetic variation resources built
by the 1000 Genomes Project. We also update the resources to the
current reference assembly, add new data sets generated from the 1000
Genomes Project samples and add data from projects working with other
openly consented samples.
}
