% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_world.R
\name{hello_world}
\alias{hello_world}
\title{Hello world}
\usage{
hello_world()
}
\value{
A string "Data science is fantastic!".
}
\description{
Hello world and add "AwesomePackage" to NAMESPACE.
}
\examples{
hello_world()
}
