% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_HGDP.R
\docType{data}
\name{map_HGDP}
\alias{map_HGDP}
\title{Corresponding tables for individuals and populations of the HGDP dataset}
\format{
\code{map_HGDP} is a table with three columns.
\describe{
\item{\code{Superpop}}{Large populations, such as Asians, Africans, Europeans.}
\item{\code{Pop}}{Small populations, such as Chinese, British, Norwegian.}
\item{\code{Indiv}}{Individuals.}}
}
\description{
\code{map_HGDP} gives the corresponding relationship among individuals,
small populations and large populations. \code{plot_structure} uses this
relationship to group data and draw structure plot.
}
\seealso{
\code{plot_structure}
}
