% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd_fit_em.R
\name{psd_fit_em}
\alias{psd_fit_em}
\title{Use EM algorithm to fit PSD model}
\usage{
psd_fit_em(G, K, maxiter)
}
\arguments{
\item{G}{The I x J matrix of counts; all entries of G should
be taken from \{0,1,2\}.}

\item{K}{An integer 2 or greater giving the matrix rank.}

\item{maxiter}{The maximum number of iterations.}
}
\value{
A matrix represents the proportion of population distribution.
}
\description{
Fit PSD model with EM algorithm, and use the loss
function as a stopping criterion.
}
\examples{
G <- matrix(c(0,0,1,0,2, 1,0,0,1,0, 0,1,0,0,1), nrow = 3)
psd_fit_em(G, 2, 10)
}
