% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd_fit_vi.R
\name{psd_fit_vi}
\alias{psd_fit_vi}
\title{Use VI algorithm to fit PSD model}
\usage{
psd_fit_vi(G, K, epsilon = 0.1, maxiter = 500)
}
\arguments{
\item{G}{The I x J matrix of counts; all entries of G should
be taken from \{0,1,2\}.}

\item{K}{An integer 2 or greater giving the matrix rank.}

\item{epsilon}{Convergence criterion.}

\item{maxiter}{The maximum number of iterations.}
}
\value{
A List with the following parameters:
\describe{
\item{\code{P}}{The population scale matrix of the individuals.}
\item{\code{F}}{The gene scale matrix of the populations.}
\item{\code{Loss}}{A vector represents the value of the loss function which records once for ten iterations.}
\item{\code{Iterations}}{An integer represents the number of iterations.}}
}
\description{
Fit PSD model with VI algorithm, and use the loss
function as a stopping criterion.
}
\examples{
G <- matrix(c(0,0,1, 0,2,1, 1,0,1, 0,1,0, 1,0,0), 3, 5)
psd_fit_vi(G, 2, 1e-1, 10)
}
