% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd_evaluate.R
\name{psd_loglikelihood}
\alias{psd_loglikelihood}
\title{Compute the loglikelihood}
\usage{
psd_loglikelihood(G, result)
}
\arguments{
\item{G}{The I x J matrix of counts; all entries of G should
be taken from \{0,1,2\}.}

\item{result}{Output of \code{psd_fit_em}, \code{psd_fit_sqp}, or \code{psd_fit_vi}.}
}
\value{
A value indicates the accuracy of the prediction.
}
\description{
Compute the maximum loglikelihood function of the PSD model.
}
\examples{
G <- matrix(c(0,0,1, 0,2,1, 1,0,1, 0,1,0, 1,0,0), 3, 5)
result <- psd_fit_em(G, 2, 1e-5, 10)
psd_loglikelihood(G, result)
}
