% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd_evaluate.R
\name{psd_loglikelihood_svi}
\alias{psd_loglikelihood_svi}
\title{Compute the loglikelihood of svi method in validation set}
\usage{
psd_loglikelihood_svi(G, K, P, maxiter = 2000)
}
\arguments{
\item{G}{The I x J matrix of counts; all entries of G should
be taken from \{0,1,2\}.}

\item{K}{An integer 2 or greater giving the matrix rank.}

\item{P}{The \code{P} in the output list of \code{psd_fit_svi}.}

\item{maxiter}{The maximum number of iterations in validation set.}
}
\value{
A value indicates the accuracy of the prediction in validation set.
}
\description{
Compute the maximum loglikelihood function of the PSD model in validation set using svi method.
}
\examples{
# Refer to Articles in AwesomePackage.
}
