% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_em.R
\docType{data}
\name{result_em}
\alias{result_em}
\title{Pre-training result of genome_chr1 by EM algorithm}
\format{
\code{result_em} is a list with four elements.
}
\description{
This is the result of the following code: \code{psd_fit_em(genome_chr1, 3, 1e-3, 500)}.
}
\seealso{
\code{psd_fit_em}
}
