% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_HGDP_vi.R
\docType{data}
\name{result_HGDP_vi}
\alias{result_HGDP_vi}
\title{Pre-training result of data_HGDP by VI algorithm}
\format{
\code{result_HGDP_vi} is a list with four elements.
}
\description{
This is the result of the following code: \code{psd_fit_vi(data_HGDP, 7, 1e-5, 800)}.
}
\seealso{
\code{psd_fit_vi}
}
