% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_sqp.R
\docType{data}
\name{result_sqp}
\alias{result_sqp}
\title{Pre-training result of genome_chr1 by SQP algorithm}
\format{
\code{result_sqp} is a list with four elements.
}
\description{
This is the result of the following code: \code{psd_fit_sqp(genome_chr1, 3, 1e-3, 500, 100)}.
}
\seealso{
\code{psd_fit_sqp}
}
