% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_TGP_em.R
\docType{data}
\name{result_TGP_em}
\alias{result_TGP_em}
\title{Pre-training result of data_TGP by EM algorithm}
\format{
\code{result_TGP_em} is a list with four elements.
}
\description{
This is the result of the following code: \code{psd_fit_em(data_TGP, 3, 1e-5, 500)}.
}
\seealso{
\code{psd_fit_em}
}
