% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_TGP_sqp.R
\docType{data}
\name{result_TGP_sqp}
\alias{result_TGP_sqp}
\title{Pre-training result of data_TGP by SQP algorithm}
\format{
\code{result_TGP_sqp} is a list with four elements.
}
\description{
This is the result of the following code: \code{psd_fit_sqp(data_TGP, 3, 1e-5, 50, 100)}.
}
\seealso{
\code{psd_fit_sqp}
}
