% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_TGP_svi.R
\docType{data}
\name{result_TGP_svi}
\alias{result_TGP_svi}
\title{Pre-training result of data_TGP by SVI algorithm}
\format{
\code{result_TGP_svi} is a list with three elements.
}
\description{
This is the result of the following code: \code{psd_fit_svi(data_TGP, 3, 1e-5, 5e+5, 100, 5e-2, 1e-1, 1e+4, 1, 0.5)}.
}
\seealso{
\code{psd_fit_svi}
}
