% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_TGP_vi.R
\docType{data}
\name{result_TGP_vi}
\alias{result_TGP_vi}
\title{Pre-training result of data_TGP by VI algorithm}
\format{
\code{result_TGP_vi} is a list with four elements.
}
\description{
This is the result of the following code: \code{psd_fit_vi(data_TGP, 3, 1e-5, 500)}.
}
\seealso{
\code{psd_fit_vi}
}
