% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_vi.R
\docType{data}
\name{result_vi}
\alias{result_vi}
\title{Pre-training result of genome_chr1 by VI algorithm}
\format{
\code{result_vi} is a list with four elements.
}
\description{
This is the result of the following code: \code{psd_fit_vi(genome_chr1, 3, 1e-9, 500)}.
}
\seealso{
\code{psd_fit_vi}
}
