% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_plot.R
\name{structure_plot}
\alias{structure_plot}
\title{Plot the ancestral proportions}
\usage{
structure_plot(L, topics, colors, ticks = NULL, font.size = 9)
}
\arguments{
\item{L}{The proportion matrix.}

\item{topics}{Topic order options.}

\item{colors}{Theme color options.}

\item{ticks}{Grouping options.}

\item{font.size}{Font size used in plot.}
}
\description{
Plot the ancestral proportions of individuals using package ggplot2.
}
\examples{
P <- matrix(c(0.5,0.5, 0.3,0.7, 0.8,0.2), 3, 2)
topics <- order(colMeans(P))
colors <- c("red", "yellow")
structure_plot(P, topics, colors)
}
