% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_structure.R
\name{plot_structure}
\alias{plot_structure}
\title{Plot the population proportion}
\usage{
plot_structure(P, pops, colors, ticks = NULL, font.size = 9)
}
\arguments{
\item{P}{The proportion matrix.}

\item{pops}{Population order options.}

\item{colors}{Theme color options.}

\item{ticks}{Grouping options.}

\item{font.size}{Font size used in plot.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot the population proportion of individuals using package ggplot2.
}
\examples{
P <- matrix(c(0.5,0.3,0.8, 0.5,0.7,0.2), 3, 2)
pops <- order(colMeans(P))
colors <- c("red", "yellow")
plot_structure(P, pops, colors)
}
